+++ examples/distributed-computing/dispatcher.py
@@ -5,7 +5,7 @@ try:
 except ImportError:
     import Queue as queue
 import Pyro4
-from Pyro4.util import SerializerBase
+from Pyro4.serializers import SerializerBase
 from workitem import Workitem
