+++ examples/distributed-computing/worker.py
@@ -4,7 +4,7 @@ import socket
 import sys
 from math import sqrt
 import Pyro4
-from Pyro4.util import SerializerBase
+from Pyro4.serializers import SerializerBase
 from workitem import Workitem
